moved find completed steps to completed steps model
