+++ src/model/index.js
@@ -1,13 +1,13 @@
 const { findRequirementsByTrack } = require("./requirements")
 const { findAllTracks } = require("./tracks")
 const { findUsers, findUsersBy, insertUser } = require("./users")
-const {
-    findStepsByTask,
-    findCompletedStepsBy,
-    getFormattedSteps
-} = require("./steps")
+const { findStepsByTask, getFormattedSteps } = require("./steps")
 
-const { markComplete, markIncomplete } = require("./completedSteps")
+const {
+    findCompletedStepsBy,
+    markComplete,
+    markIncomplete
+} = require("./completedSteps")
 
 module.exports = {
     findUsers,
