+++ tests/e2e/completedSteps-model.test.js
@@ -30,7 +30,34 @@ describe("MODEL steps", () => {
         await db.destroy()
         done()
     })
-
+    describe("findCompletedStepsBy", () => {
+        it("should return an array of objects", done => {
+            findCompletedStepsBy({ user_id: 1 }).then(res => {
+                expect(res).toEqual(expect.any(Array))
+                expect(res).toContainEqual(expect.any(Object))
+                done()
+            })
+        })
+        it("should return an array of objects with shape: {id (int), user_id (int), steps_id (bool), created_at (String)} ", done => {
+            findCompletedStepsBy({ user_id: 1 }).then(res => {
+                expect(res[0]).toEqual(
+                    expect.objectContaining({
+                        id: expect.any(Number),
+                        user_id: expect.any(Number),
+                        steps_id: expect.any(Number),
+                        created_at: expect.any(Date)
+                    })
+                )
+                done()
+            })
+        })
+        it("should have this object as it's first element: { user_id: 1, steps_id: 1 },", done => {
+            findCompletedStepsBy({ user_id: 1 }).then(res => {
+                expect(res[0]).toEqual({ ...fakeCompletedSteps[0], id: 1 })
+                done()
+            })
+        })
+    })
     describe("markComplete", () => {
         it("should return an array with an object", done => {
             markComplete(1, 3).then(res => {
